app/requirements.txt
psycopg[binary]==3.2.1
Faker==17.6.0
python-dotenv==1.0.1
